[NUM-24] Created a new page for login
