+++ src/App.js
@@ -3,6 +3,7 @@ import Home from './pages/Home';
 import './App.css';
 import store from './store/store'
 import {Provider} from "react-redux";
+import Login from './pages/Login';
 
 function App() {
 
@@ -11,9 +12,14 @@ function App() {
           <BrowserRouter>
               <Routes>
                   <Route path='/' element={<Home />} />
+                  <Route path='/Login' element={<Login/>} />
               </Routes>
           </BrowserRouter>
       </Provider>
   );
 }
 export default App;
+
+
+
+
