+++ src/components/Register/register.module.css
@@ -1,0 +1,12 @@
+.form{
+    height: 44px;
+    width: 332px;
+    margin-left: 46px;
+    border-radius: 6px;
+    align-items: center;
+}
+
+.buton{
+    padding-left: 110px;
+    padding-right: 100px;
+}
