+++ src/components/index.js
@@ -6,3 +6,5 @@ export { default as Category } from './Category/Category'
 export { default as DestinationTitle } from './DestinationTitle/DestinationTitle'
 export { default as DestinationPrice } from './DestinationPrice/DestinationPrice'
 export { default as DestinationCard } from './DestinationCard/DestinationCard'
+export { default as Register } from './Register/Register'
+export {default as LoginRegister } from './LoginRegister/LoginRegister'
