+++ src/pages/Login.js
@@ -1,0 +1,11 @@
+import { LoginRegister, Register,  } from "../components"
+
+
+function Login (){
+    return(
+        <div>
+           <LoginRegister/>
+        </div>
+    )
+}
+export default Login
